+++ src/core/Table/Table.tsx
@@ -134,6 +134,11 @@ export type TableProps<
    * Must be memoized.
    */
   rowProps?: (row: Row<T>) => React.ComponentPropsWithRef<'div'>;
+  /**
+   * Modify the density of the table (adjusts the row height).
+   * @default 'default'
+   */
+  density?: 'default' | 'condensed' | 'extra-condensed';
 } & Omit<CommonProps, 'title'>;
 
 /**
@@ -207,6 +212,7 @@ export const Table = <
     expanderCell,
     isRowDisabled,
     rowProps,
+    density = 'default',
     ...rest
   } = props;
 
@@ -337,7 +343,11 @@ export const Table = <
       ref={(element) => setOwnerDocument(element?.ownerDocument)}
       id={id}
       {...getTableProps({
-        className: cx('iui-table', className),
+        className: cx(
+          'iui-table',
+          { [`iui-${density}`]: density !== 'default' },
+          className,
+        ),
         style,
       })}
       {...ariaDataAttributes}